fix: Moved calculate to separate class
